+++ R/mc_miscellanea.R
@@ -5,7 +5,8 @@
 #' Counts the number of clause units (bounded by the \code{<##>}, \code{<#>}, or
 #' \code{<\%>} annotation symbols) in a multicastR table.
 #'
-#' @seealso \code{\link{multicast}}
+#' @seealso \code{\link{multicast}}, \code{\link{mc_index}},
+#'   \code{\link{mc_referents}}, \code{\link{mc_metadata}}
 #'
 #' @param mcdata A \code{\link[data.table]{data.table}} in multicastR format,
 #'   containing minimally a \code{corpus} column with the names of the corpora
@@ -22,10 +23,10 @@
 #' \dontrun{
 #'   # count clause units in the most recent version
 #'   # of the Multi-CAST data, by corpus
-#'   n <- mc_count_clauses(multicast())
+#'   n <- mc_clauses(multicast())
 #'
 #'   # count by text instead
-#'   m <- mc_count_clauses(multicast(), bytext = TRUE)
+#'   m <- mc_clauses(multicast(), bytext = TRUE)
 #'
 #'   # number of clauses units in the whole collection
 #'   sum(n$nClauses)
@@ -72,26 +73,3 @@ if (getRversion() >= "2.15.1") {
 }
 
 # ----------------------------------------------------------------------
-
-# ----------------------------------------------------------------------
-
-#' Count clause units in a multicastR table (WIP)
-#'
-#' Deprecated with multicastR 1.2.0. Use \code{\link{mc_clauses}} instead.
-#'
-#' @param mcdata A \code{\link[data.table]{data.table}} in multicastR format,
-#'   containing minimally a \code{corpus} column with the names of the corpora
-#'   and a \code{graid} column with GRAID annotation values.
-#' @param bytext Logical. If \code{FALSE}, calculate the number of clause units
-#'   for each corpus. If \code{TRUE}, count for each text separately.
-#'
-#' @return A \code{\link[data.table]{data.table}} with the number of valid
-#'   clause units in each corpus, the total number of clause units, the number
-#'   of non-analyzed clause units ("NC"), and the percentage the later make up
-#'   of the total.
-#'
-#' @export
-mc_count_clauses <- function(mcdata, bytext = FALSE) {
-	.Deprecated("mc_clauses")
-	mc_clauses(mcdata, bytext)
-}