+++ R/multicast.R
@@ -6,8 +6,9 @@
 #' the University of Bamberg and outputs them as a
 #' \code{\link[data.table]{data.table}}. As the Multi-CAST collection is
 #' amenable to extension by additional data sets and annotation schemes,
-#' \code{multicast} takes an optional argument to select earlier versions of the
-#' annotation data to ensure scientific accountability and reproducability.
+#' \code{multicast} takes an optional argument \code{vkey} to select earlier
+#' versions of the annotation data to ensure scientific accountability and
+#' reproducibility.
 #'
 #' @section Licensing: The Multi-CAST annotation data accessed by the
 #'   \code{multicast} method is published under a \emph{Create Commons
@@ -17,11 +18,11 @@
 #'   to its contributors.
 #'
 #' @section Citing Multi-CAST: Data from the Multi-CAST collection should be
-#'   cited as: \itemize{ \item Haig, Geoffrey & Schnell, Stefan (eds.).
-#'   2015. \emph{Multi-CAST: Multilinguial Corpus of Annotated Spoken
-#'   Texts}. (\url{https://multicast.aspra.uni-bamberg.de/}) (Accessed
-#'   \emph{date}.) } If for some reason you need to cite this package on its
-#'   own, please refer to \code{citation(multicastR)}.
+#'   cited as: \itemize{ \item Haig, Geoffrey & Schnell, Stefan (eds.). 2015.
+#'   \emph{Multi-CAST: Multilingual Corpus of Annotated Spoken Texts}.
+#'   (\url{https://multicast.aspra.uni-bamberg.de/}) (Accessed \emph{date}.) }
+#'   If you need to cite this package specifically, please refer to
+#'   \code{citation(multicastR)}.
 #'
 #' @section References: \itemize{\item Haig, Geoffrey & Schnell, Stefan. 2014.
 #'   \emph{Annotations using GRAID (Grammatical Relations and Animacy in
@@ -37,18 +38,14 @@
 #'   guidelines.} Version 1.1.
 #'   (\url{https://multicast.aspra.uni-bamberg.de/#annotations})}
 #'
-#' @seealso \code{\link{mc_index}}, \code{\link{mc_referents}}
+#' @seealso \code{\link{mc_index}}, \code{\link{mc_referents}},
+#'   \code{\link{mc_metadata}}, \code{\link{mc_clauses}}
 #'
 #' @param vkey A numeric or character vector of length 1 specifying the
 #'   requested version of the annotation values. Must be one of the four-digit
 #'   version keys in the first column of \code{\link{mc_index}}, or empty. If
-#'   empty or no value is supplied, \code{multicast} automatically retrieves the
-#'   most recent version of the annotations. See the examples below for an
-#'   illustration.
-#' @param legacy.colnames If \code{TRUE}, renames the \code{text} and
-#'   \code{gword} columns to what they were called prior to version 1.1.0 of the
-#'   package (i.e. \code{file}, \code{word}). This option will be removed in the
-#'   future.
+#'   empty or no value is supplied, the most recent version of the annotations
+#'   is retrieved automatically. See the examples below for an illustration.
 #'
 #' @return A \code{\link[data.table]{data.table}} with eleven columns:
 #'   \describe{ \item{\code{[, 1] corpus}}{The name of the corpus.}
@@ -64,9 +61,9 @@
 #'   gloss.} \item{\code{[, 8] ganim}}{The person-animacy symbol of a GRAID
 #'   gloss.} \item{\code{[, 9] gfunc}}{The function symbol of a GRAID gloss.}
 #'   \item{\code{[, 10] refind}}{Referent tracking using the RefIND scheme
-#'   (Schiborr et al. 2018).} \item{\code{[, 11] reflex}}{The information status
-#'   of newly introduced referents, using a simplified version of the RefLex
-#'   scheme (Riester & Baumann 2017).} }
+#'   (Schiborr et al. 2018).} \item{\code{[, 11] isnref}}{Annotations of the
+#'   information status of newly introduced referents with ISNRef, a simplified
+#'   version of the RefLex scheme (Riester & Baumann 2017).} }
 #'
 #' @examples
 #' \dontrun{
@@ -74,13 +71,14 @@
 #'   # Multi-CAST annotations
 #'   multicast()
 #'
-#'   # retrieve and print the version of the annotation data
-#'   # published in May 2019
+#'   # retrieve the version of the annotation data published
+#'   # in May 2019
 #'   multicast(1905)   # or: multicast("1905")
 #' }
 #'
+#' @importFrom curl curl
 #' @export
-multicast <- function(vkey, legacy.colnames = FALSE) {
+multicast <- function(vkey) {
 	# check whether vkey is missing
 	if (!mc_missarg(vkey)) {
 		# 1A: vkey is not missing
@@ -148,9 +146,9 @@ multicast <- function(vkey, legacy.colnames = FALSE) {
 	}
 
 	# construct URL for annotation file
-	path <- paste0("https://multicast.aspra.uni-bamberg.de/data/mcr/multicast_",
+	path <- paste0("https://multicast.aspra.uni-bamberg.de/data/",
 				   vkey,
-				   ".tsv")
+				   "/multicast.tsv")
 
 	# fetch annotation file
 	message("Retrieving annotations...")
@@ -166,11 +164,7 @@ multicast <- function(vkey, legacy.colnames = FALSE) {
 		error=function(e) { stop("failed to download annotations.") }
 	)
 	size <- index[version == vkey, 3]
-	message(paste0("Success! Downloaded ", size, "."))
-
-	if (legacy.colnames == TRUE) {
-		colnames(c("text", "gword"), c("file", "gword"), mc)
-	}
+	message(paste0("Downloaded ", size, "."))
 
 	# return annotation values
 	return(mc)
@@ -217,20 +211,10 @@ mc_index <- function() {
 									   showProgress = FALSE)),
 		error=function(e) { stop("failed to download index.") }
 	)
-	message(paste0("Success! Downloaded <1 KB."))
+	message(paste0("Downloaded <1 KB."))
 
 	# return index
 	return(index)
 }
 
 # ----------------------------------------------------------------------
-
-#' Access the Multi-CAST version index
-#'
-#' Deprecated with multicastR 1.1.0. Use \code{\link{mc_index}} instead.
-#'
-#' @export
-mcindex <- function() {
-	.Deprecated("mc_index")
-	mc_index()
-}